% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_ejson_map}
\alias{plot_ejson_map}
\title{Plot a map from EJSON}
\usage{
plot_ejson_map(ejson)
}
\arguments{
\item{ejson}{Path to file or object returned by read_ejson()}
}
\value{
A ggplot object
}
\description{
Plot a map from EJSON
}
